I added MentionEntityType to replace EntityType and changes all the other factorie code to account for this.  I moved the original rule-based EntityType predictor to the same file as the new one.
