+++ src/main/scala/cc/factorie/app/nlp/coref/WithinDocCoref2.scala
@@ -20,7 +20,7 @@ abstract class BaseWithinDocCoref2 extends DocumentAnnotator {
   val options = new Coref2Options
   val model: PairwiseCorefModel
 
-  def prereqAttrs = Seq(classOf[MentionList], classOf[EntityType], classOf[MentionGenderLabel], classOf[MentionNumberLabel])
+  def prereqAttrs = Seq(classOf[MentionList], classOf[MentionEntityType], classOf[MentionGenderLabel], classOf[MentionNumberLabel])
   def postAttrs = Seq(classOf[GenericEntityMap[Mention]])
   def process1(document: Document) = {
     if (options.useEntityLR) document.attr += processDocumentOneModelFromEntities(document)
