+++ src/main/scala/cc/factorie/app/nlp/coref/WithinDocCoref2Trainer.scala
@@ -2,7 +2,7 @@ package cc.factorie.app.nlp.coref
 
 import cc.factorie.app.nlp.Document
 import cc.factorie.util.coref.GenericEntityMap
-import cc.factorie.app.nlp.mention.{MentionNumberLabeler, MentionGenderLabeler, MentionList, Mention}
+import cc.factorie.app.nlp.mention._
 import cc.factorie.app.nlp.wordnet.WordNet
 import java.io.File
 
@@ -194,6 +194,9 @@ object WithinDocCoref2Trainer {
     import cc.factorie.app.nlp.Implicits._
     val (trainDocs,trainMap) = if(loadTrain) MentionAlignment.makeLabeledData(trainFile,null,opts.portion.value,options.useEntityType, options) else (null,null)
     val (testDocs,testMap) = MentionAlignment.makeLabeledData(testFile,null,opts.portion.value,options.useEntityType, options)
+    trainDocs.foreach(MentionEntityTypeLabeler.process1(_))
+    testDocs.foreach(MentionEntityTypeLabeler.process1(_))
+
     (trainDocs,trainMap,testDocs,testMap)
   }
 }
