+++ src/main/scala/cc/factorie/app/nlp/mention/MentionType.scala
@@ -2,16 +2,20 @@ package cc.factorie.app.nlp.mention
 import cc.factorie._
 import cc.factorie.app.nlp._
 import cc.factorie.app.nlp.pos._
-import cc.factorie.app.nlp.coref.EntityType
 import cc.factorie.app.nlp.ner.OntonotesNerDomain
 import cc.factorie.util.BinarySerializer
 import cc.factorie.optimize._
 import java.io._
 
-// TODO EntityType should be moved here, and renamed MentionType (since it has a "val mention" among its arguments). -akm
+//'Entity Type' is a misnomer that is used elsewhere in the literature, use it too. Really, this is a type associated with a mention, not an entity
 
-/** Cheap number predictor based on rules and lexicons.  Really this should use a real morphological analyzer. */
-class MentionTypeLabeler extends DocumentAnnotator {
+
+class MentionEntityType(val mention:Mention, targetValue:String) extends LabeledCategoricalVariable(targetValue) {
+  def domain = OntonotesNerDomain
+}
+
+
+class MentionEntityTypeLabeler extends DocumentAnnotator {
   def this(stream:InputStream) = { this(); deserialize(stream) }
   def this(file: File) = this(new FileInputStream(file))
   def this(url:java.net.URL) = this(url.openConnection.getInputStream)
@@ -43,13 +47,13 @@ class MentionTypeLabeler extends DocumentAnnotator {
   def process1(document:Document): Document = {
     import MentionNumberDomain._
     for (mention <- document.attr[MentionList])
-      mention.attr.getOrElseUpdate(new EntityType(mention, "O")) := model.classification(features(mention).value).bestLabelIndex
+      mention.attr.getOrElseUpdate(new MentionEntityType(mention, "O")) := model.classification(features(mention).value).bestLabelIndex
     document
   }
-  override def tokenAnnotationString(token:Token): String = { val mentions = token.document.attr[MentionList].filter(_.span.contains(token)); mentions.map(_.attr[EntityType].categoryValue).mkString(",") }
-  override def mentionAnnotationString(mention:Mention): String = { val t = mention.attr[EntityType]; if (t ne null) t.categoryValue else "_" }
+  override def tokenAnnotationString(token:Token): String = { val mentions = token.document.attr[MentionList].filter(_.span.contains(token)); mentions.map(_.attr[MentionEntityType].categoryValue).mkString(",") }
+  override def mentionAnnotationString(mention:Mention): String = { val t = mention.attr[MentionEntityType]; if (t ne null) t.categoryValue else "_" }
   def prereqAttrs: Iterable[Class[_]] = List(classOf[MentionList])
-  def postAttrs: Iterable[Class[_]] = List(classOf[EntityType])
+  def postAttrs: Iterable[Class[_]] = List(classOf[MentionEntityType])
  
   def filterMentions(mentions:Seq[Mention]): Iterable[Mention] = mentions.groupBy(m => m.attr[Entity]).filter(x => x._2.length > 1).map(x => x._2).flatten
 
@@ -61,12 +65,12 @@ class MentionTypeLabeler extends DocumentAnnotator {
     FeatureDomain.dimensionDomain.trimBelowCount(3)
     //val entityMentions = allMentions.groupBy(m => m.attr[Entity]).filter(x => x._2.length > 1).map(x => x._2).flatten
     val examples = for (doc <- trainDocs; mention <- filterMentions(doc.attr[MentionList])) yield
-      new LinearMultiClassExample(model.weights, features(mention).value, mention.attr[EntityType].intValue, LinearObjectives.hingeMultiClass)
+      new LinearMultiClassExample(model.weights, features(mention).value, mention.attr[MentionEntityType].intValue, LinearObjectives.hingeMultiClass)
     val testMentions = testDocs.flatMap(doc => filterMentions(doc.attr[MentionList]))
     println("Training ")
     def evaluate(): Unit = {
-      println("TRAIN\n"+(new cc.factorie.app.classify.Trial[EntityType,la.Tensor1](model, OntonotesNerDomain, (t:EntityType) => features(t.mention).value) ++= trainMentions.map(_.attr[EntityType])).toString)
-      println("\nTEST\n"+(new cc.factorie.app.classify.Trial[EntityType,la.Tensor1](model, OntonotesNerDomain, (t:EntityType) => features(t.mention).value) ++= testMentions.map(_.attr[EntityType])).toString)
+      println("TRAIN\n"+(new cc.factorie.app.classify.Trial[MentionEntityType,la.Tensor1](model, OntonotesNerDomain, (t:MentionEntityType) => features(t.mention).value) ++= trainMentions.map(_.attr[MentionEntityType])).toString)
+      println("\nTEST\n"+(new cc.factorie.app.classify.Trial[MentionEntityType,la.Tensor1](model, OntonotesNerDomain, (t:MentionEntityType) => features(t.mention).value) ++= testMentions.map(_.attr[MentionEntityType])).toString)
     }
     Trainer.onlineTrain(model.parameters, examples.toSeq, maxIterations=3, evaluate = evaluate)
   }
@@ -101,23 +105,119 @@ class MentionTypeLabeler extends DocumentAnnotator {
 
 }
 
-object MentionTypeLabeler extends MentionTypeLabeler(cc.factorie.util.ClasspathURL[MentionTypeLabeler](".factorie"))
+object MentionEntityTypeLabeler extends MentionEntityTypeLabeler(cc.factorie.util.ClasspathURL[MentionEntityTypeLabeler](".factorie"))
 
-object MentionTypeLabelerTrainer {
+object MentionEntityTypeLabelerTrainer {
   def main(args:Array[String]): Unit = {
     println("usage: trainfile [modelfile]")
     var trainDocs = coref.ConllCorefLoader.loadWithParse(args(0), loadSingletons=false, disperseEntityTypes=true)
     val testDocs = trainDocs.takeRight(20)
-    val labeler = new MentionTypeLabeler
+    val labeler = new MentionEntityTypeLabeler
     for (mention <- labeler.filterMentions(testDocs.flatMap(_.attr[MentionList])))
-      println("%20s  %s".format(mention.attr[EntityType].target.categoryValue, mention.span.phrase))
+      println("%20s  %s".format(mention.attr[MentionEntityType].target.categoryValue, mention.span.phrase))
 
     trainDocs = trainDocs.dropRight(20)
     labeler.train(trainDocs, testDocs)
     for (mention <- labeler.filterMentions(testDocs.flatMap(_.attr[MentionList])))
-      println("%20s %-20s  %s".format(mention.attr[EntityType].target.categoryValue, mention.attr[EntityType].categoryValue, mention.span.phrase))
+      println("%20s %-20s  %s".format(mention.attr[MentionEntityType].target.categoryValue, mention.attr[MentionEntityType].categoryValue, mention.span.phrase))
 
     if (args.length > 1) labeler.serialize(args(1))
     
   }
+}
+
+
+//this gives each Mention and MentionEntityType. This is a very simple rule-based annotator that can not even produce predictions
+//for many of the categories in  OntonotesMentionEntityTypeDomain, only PERSON, ORG, GPE, and EVENT.
+object MentionEntityTypeAnnotator1 extends DocumentAnnotator {
+  import MentionEntityTypeAnnotator1Util._
+  def process1(document:Document): Document = {
+    document.attr[MentionList].foreach(predictMentionEntityType(_))
+    document
+  }
+  def predictMentionEntityType(m: Mention): Unit = {
+    val prediction = classifyUsingRules(m.span.tokens.map(_.lemmaString))
+    m.attr += new MentionEntityType(m,prediction)
+  }
+  override def tokenAnnotationString(token:Token): String = {
+    token.document.attr[MentionList].filter(mention => mention.span.contains(token)) match { case ms:Seq[Mention] if ms.length > 0 => ms.map(m => m.attr[MentionEntityType].categoryValue + ":" + m.span.indexOf(token)).mkString(","); case _ => "_" }
+  }
+  def prereqAttrs: Iterable[Class[_]] = List(classOf[MentionList])
+  def postAttrs: Iterable[Class[_]] = List(classOf[MentionEntityType])
+
+}
+
+//the reason some functions for above are pulled into a separate object as this simplifies the API with other projects greatly
+object MentionEntityTypeAnnotator1Util {
+  final val articles = Seq("a","A","the","The").toSet
+
+  //this expects cased strings as input
+  def classifyUsingRules(strings: Seq[String]): String = {
+
+    val uStr = strings.filter(!articles.contains(_))
+    val str = uStr.map(_.toLowerCase)
+    val str1 = strings.mkString(" ")
+    val uStr1 = uStr.mkString(" ")
+
+    val isPerson = detectIfPerson(str,uStr)
+    val isPlace = detectIfPlace(str1,uStr1)
+    val isEvent = detectIfEvent(str1,uStr1)
+    val isOrganization = detectIfOrg(str1,uStr1)
+    val onlyOne =  Seq(isPerson, isPlace, isEvent,  isOrganization).count(y => y) == 1
+
+    if(onlyOne){
+      if(isPerson) return "PERSON"
+      else if(isPlace) return "GPE"
+      else if(isEvent) return "EVENT"
+      else if(isOrganization) return "ORG"
+      else
+        return "O"
+    }else{
+      if(isPlace && isOrganization) //the place lexicon is mostly contained in the organization lexicon, so you need to treat it carefully.
+        return "GPE"
+      else if(isPlace && isPerson)
+        return "GPE"
+      else
+        return "O"
+    }
+  }
+
+  def detectIfPerson(strs: Seq[String], uStrs: Seq[String]): Boolean = {
+    val isCased = strs.zip(uStrs).exists(ab => ab._1 != ab._2)
+    val str = strs.mkString(" ")
+    val uStr = uStrs.mkString(" ")
+    val fullStringPerson = lexicon.wikipedia.Person.contains(str) && isCased
+    val fields = strs
+    val uFields = uStrs
+
+    val firstIsCased = fields.nonEmpty && fields(0) != uFields(0)
+    val secondIsCased = if(fields.length == 2) fields(1) != uFields(1) else false
+    val firstContained = fields.nonEmpty && (lexicon.uscensus.PersonFirstFemale.contains(fields(0)) || lexicon.uscensus.PersonFirstMale.contains(fields(0)))
+    val secondContained =   if(fields.length == 2) lexicon.uscensus.PersonLast.contains(fields(1))  else false
+
+    val firstName = fields.length == 2  && firstContained  &&  firstIsCased  && secondIsCased
+    val firstName2 = fields.length == 1 && lexicon.iesl.PersonFirst.contains(fields(0))  &&  firstIsCased
+    val lastName = fields.length == 2   && firstContained && secondContained   &&  secondIsCased && firstIsCased
+    val lastName2 = fields.length == 1   && lexicon.uscensus.PersonLast.contains(fields(0))   &&  firstIsCased
+    val bothNames = fields.length == 2 && firstContained && secondContained
+    val isI = fields.length == 1 && uStr == "I"
+
+    val isPerson = lastName || lastName2 || firstName || firstName2|| fullStringPerson  || bothNames
+    isPerson  && ! isI
+  }
+  //the follow three methods just check for exact string matches
+  def detectIfPlace(s: String, us: String): Boolean = {
+    lexicon.iesl.AllPlaces.contains(s)
+  }
+  def detectIfOrg(s: String, us: String): Boolean = {
+    lexicon.iesl.OrgSuffix.contains(s)
+  }
+  def detectIfEvent(s: String, us: String): Boolean = {
+    lexicon.wikipedia.Event.contains(s)
+  }
+
+  def classifyUsingRules(rawString: String): String = {
+    classifyUsingRules(rawString.split(" "))
+  }
+
 }
